==> src/12_explore_DVA_results_sex_stratified.Rmd <==
---
title: "09_explore_results_sex_Stratified"
author: "Rogério Ribeiro"
date: "`r Sys.Date()`"
output: html_document
---

# Explore Results from the scRNA differential variability analysis

# Load libraries
```{r}
shh <- suppressPackageStartupMessages

shh(library(tidyverse))
shh(library(ggrastr))
shh(library(reshape2))
shh(library(ComplexHeatmap))
shh(library(RColorBrewer))
shh(library(gridExtra))
shh(library(ggrepel))
```


# Set image dir
```{r}
main.dir <-  "../../results/azimuth_annotation_results/"
img.dir <- "../../results/azimuth_annotation_results/10_explore_DVA_results_sex_stratified/"
if (!dir.exists(img.dir)) {dir.create(img.dir)}
```


# Load data
```{r}
cts <- c("Bintermediate", "Bmemory", "Bnaive", "CD4CTL", "CD4Naive", 
                "CD4Proliferating", "CD4TCM", "CD4TEM", "CD8Naive", "CD8Proliferating", 
                "CD8TCM", "CD8TEM", "CD14Mono", "CD16Mono", "cDC2", "dnT", "Eryth", 
                "gdT", "HSPC", "MAIT", "NK_CD56bright", "NK", "NKProliferating", 
                "pDC", "Plasmablast", "Platelet", "Treg")


cts_to_study <- c("Bintermediate", "Bmemory", "Bnaive", "CD4CTL", "CD4Naive", "CD4TCM", "CD4TEM", "CD8Naive", 
                "CD8TCM", "CD8TEM", "CD14Mono", "CD16Mono","MAIT", "NK_CD56bright", "NK", "Treg")


#cts_to_study <- c("CD4_NC", "CD8_ET", "CD8_NC", "NK")
```

## Load mean expression

```{r}
mean_exps <- list()
for (ct in cts){
  mean_exps[[ct]] <- read.table(paste0(main.dir, "03_mean_variance_estimates/", ct, "_cells_mean_mx.txt"))
}

```

## Load dispersion estimates

```{r}
dispersions <- list()
for (ct in cts){
  dispersions[[ct]] <- read.table(paste0(main.dir, "04_individual_dispersions/", ct, "_within_ind_dispersion.txt"), header = TRUE)
  row.names(dispersions[[ct]]) <- dispersions[[ct]]$gene
  dispersions[[ct]]$gene <- NULL
}
```

## Load DVG list

```{r}
DVGs <- readRDS(paste0(main.dir, "06_model_fit/1.DVG_results.sex_stratified.rds")) # Open file
DVGs <- lapply(DVGs, function(x) lapply(x, function(y) y %>% mutate(significance=ifelse(p_value_adj < 0.05, "sig", "not_sig"))))

DVGs_new = list("M" = list(), 
                "Fe"  =  list()
                )

for (x in names(DVGs$M)){
  if (x %in% cts_to_study){
    DVGs_new$M[[x]] <- DVGs$M[[x]]
    DVGs_new$Fe[[x]] <- DVGs$F[[x]]
  }
}

DVGs <- DVGs_new
```

## Load sample size

```{r}
# Create the dataframe with the provided values
sample_size <- data.frame(
  celltype = c("Bintermediate", "Bmemory", "Bnaive", "CD4CTL", "CD4Naive", 
               "CD4Proliferating", "CD4TCM", "CD4TEM", "CD8Naive", "CD8Proliferating",
               "CD8TCM", "CD8TEM", "CD14Mono", "CD16Mono", "cDC2", "dnT", "Eryth",
               "gdT", "HSPC", "MAIT", "NK_CD56bright", "NK", "NKProliferating", "pDC",
               "Plasmablast", "Platelet", "Treg"),
  median = c(24, 20, 60, 7, 289, 1, 239, 21, 39, 1, 9, 150, 24, 11,    4, 3, 1, 3, 2, 9, 7, 154, 3, 2, 3, 2, 26),
  median_M = c(22, 16, 50, 7, 258, 1, 243, 19, 35, 1, 9, 155, 27, 12, 4, 3, 1, 4, 2, 8, 6, 172.5, 3, 2, 3, 2, 25),
  median_F = c(26, 22, 68, 7, 313, 1, 236, 23, 43, 1, 9, 146, 22, 10, 4, 3, 1, 3, 2, 10, 8, 141, 3, 2, 3, 2, 28)
)


sample_size <- sample_size %>% 
  filter(celltype %in% cts_to_study)
```


## Load metadata

```{r}
metadata <- readRDS("../../data/donor_metadata.rds") %>%
  mutate(sex = as.factor(Gender),
         sample_id = assignment, 
         age = Age) %>% 
  select(sample_id, sex, age)
```


# Before filtering
## Summarise the result

```{r}
# Male
number_of_genes_m <- lapply(names(DVGs$M), function(x){
    temp <- DVGs$M[[x]] %>% filter(significance == "sig")
    data.frame("celltype" = x, "up" = sum(temp$beta > 0), "down" = sum(temp$beta < 0))
  }
)

number_of_genes_M <- do.call("rbind", number_of_genes_m)
df_melted_M <- melt(number_of_genes_M, id.vars = "celltype", variable.name = "direction", value.name = "count")


# Female
number_of_genes_Fe <- lapply(names(DVGs$Fe), function(x){
    temp <- DVGs$Fe[[x]] %>% filter(significance == "sig")
    data.frame("celltype" = x, "up" = sum(temp$beta > 0), "down" = sum(temp$beta < 0))
  }
)

number_of_genes_Fe <- do.call("rbind", number_of_genes_Fe)
df_melted_Fe <- melt(number_of_genes_Fe, id.vars = "celltype", variable.name = "direction", value.name = "count")

```

## Make an heatmap

```{r, fig.height=8, fig.width=7}
col = brewer.pal(9, "BuPu")[1:7]

dat1 <- number_of_genes_M[,c(2:3)]
dat2 <- number_of_genes_Fe[,c(2:3)]

row.names(dat1) <- number_of_genes_M[,1]
row.names(dat2) <- number_of_genes_Fe[,1]


#n_counts <- sample_size$N
median_cellsM <- sample_size$median_M
median_cellsF <- sample_size$median_F
celltypes <- sample_size$celltype

dat1 <- dat1[celltypes,]
dat2 <- dat2[celltypes,]

# Create the heatmaps as you had before
row_ha_left <- rowAnnotation(#`N subjects`= anno_barplot(n_counts, 
                            #                   gp = gpar(fill = "blue"), 
                            #                  width = unit(1, "cm"),  # Decreased width for the N barplot
                            #                   border = TRUE), 
                             `Median\ncell per\nsubject` = anno_barplot(median_cellsM, 
                                                   gp = gpar(fill = "green"), 
                                                   width = unit(1, "cm"),  # Decreased width for the Median barplot
                                                   border = TRUE)
                             )

row_ha_left_f <- rowAnnotation(#`N subjects`= anno_barplot(n_counts, 
                            #                   gp = gpar(fill = "blue"), 
                            #                  width = unit(1, "cm"),  # Decreased width for the N barplot
                            #                   border = TRUE), 
                             `Median\ncell per\nsubject` = anno_barplot(median_cellsF, 
                                                   gp = gpar(fill = "green"), 
                                                   width = unit(1, "cm"),  # Decreased width for the Median barplot
                                                   border = TRUE)
                             )

heatmap1 <- Heatmap(as.matrix(dat1),
                    name = "DVG",               
                    col = brewer.pal(9,"BuPu")[1:7],  
                    cluster_rows = FALSE,      
                    cluster_columns = FALSE,    
                    show_column_names = TRUE,   
                    show_row_names = FALSE,
                    left_annotation = row_ha_left,
                    column_title = "Male",
                    cell_fun = function(j, i, x, y, width, height, fill) {
                     grid.text(prettyNum(dat1[i, j], big.mark = ","), x, y, gp = gpar(fontsize = 10))}
                    )

heatmap2 <- Heatmap(as.matrix(dat2),
                    name = "DVG",               
                    col = brewer.pal(9,"BuPu")[1:7],  
                    cluster_rows = FALSE,      
                    cluster_columns = FALSE,    
                    show_column_names = TRUE,
                    left_annotation = row_ha_left,
                    show_row_names = TRUE,
                    column_title = "Female",
                    cell_fun = function(j, i, x, y, width, height, fill) {
                     grid.text(prettyNum(dat2[i, j], big.mark = ","), x, y, gp = gpar(fontsize = 10))}
                    )





# Draw the combined heatmap with a gap between the elements
draw(heatmap1 + heatmap2 , ht_gap = unit(1, "cm"))

pdf(paste0(img.dir, "01_dvg_before_filtering.pdf"), height=6, width=5)
draw(heatmap1 + heatmap2, ht_gap = unit(1, "cm"))
dev.off()
```

## Plot histogram for p.value and betas

```{r}
complete_table_M <- do.call("rbind", lapply(names(DVGs$M), function(x) DVGs$M[[x]] %>% mutate(ct = x)))
complete_table_fe <- do.call("rbind", lapply(names(DVGs$Fe), function(x) DVGs$Fe[[x]] %>% mutate(ct = x)))

```

### P.value for Male

```{r fig.height=8, fig.width=6}
a <- ggplot(complete_table_M, aes(x = p_value)) +
  geom_histogram(fill = "lightblue", colour = "grey30", binwidth = 0.01, show.legend = FALSE) +
  geom_vline(xintercept = c(0.05), color = "red", linetype = "dashed", alpha = 0.5, size = 0.5) +
  xlab(bquote("pvalue")) +
  ylab("Counts") +
  theme_bw() +
  theme(strip.text = element_text(size = 12)) + 
  ggtitle("Male") + 
  facet_wrap(~ct, ncol = 2, scales = "free_y")


a 

pdf(paste0(img.dir, "01_pvalue.M.pdf"), height=22, width=16)
print(a)
dev.off()
```

### P.value for Female

```{r fig.height=8, fig.width=6}
a <- ggplot(complete_table_fe, aes(x = p_value)) +
  geom_histogram(fill = "lightblue", colour = "grey30", binwidth = 0.01, show.legend = FALSE) +
  geom_vline(xintercept = c(0.05), color = "red", linetype = "dashed", alpha = 0.5, size = 0.5) +
  xlab(bquote("pvalue")) +
  ylab("Counts") +
  theme_bw() +
  theme(strip.text = element_text(size = 12)) + 
  ggtitle("Sex") + 
  facet_wrap(~ct, ncol = 2, scales = "free_y")


a 

pdf(paste0(img.dir, "01_pvalue.F.pdf"), height=22, width=16)
print(a)
dev.off()
```

### Betas for Male

```{r fig.height=8, fig.width=6}
a <- ggplot(complete_table_M, aes(x = beta, fill = significance)) +
  geom_histogram(alpha = 0.5, binwidth = 0.001, show.legend = T, position = "identity") +
  geom_vline(xintercept = c(0.05), color = "red", linetype = "dashed", alpha = 0.5, size = 0.5) +
  xlab(bquote("Coefficient")) +
  ylab("Counts") +
  theme_bw() +
  theme(strip.text = element_text(size = 12)) +
  facet_wrap(~ct, ncol = 2, scales = "free_y")


a 

pdf(paste0(img.dir, "01_beta.M.pdf"), height=22, width=16)
print(a)
dev.off()
```

### Betas for sex

```{r fig.height=8, fig.width=6}
a <- ggplot(complete_table_fe, aes(x = beta, fill = significance)) +
  geom_histogram(alpha = 0.5, binwidth = 0.01, show.legend = T, position = "identity") +
  geom_vline(xintercept = c(0.05), color = "red", linetype = "dashed", alpha = 0.5, size = 0.5) +
  xlab(bquote("Coefficient")) +
  ylab("Counts") +
  theme_bw() +
  theme(strip.text = element_text(size = 12)) + 
  facet_wrap(~ct, ncol = 2, scales = "free_y")


a 

pdf(paste0(img.dir, "01_beta.F.pdf"), height=22, width=16)
print(a)
dev.off()
```

## Plot mean vs dispersion in significant genes (Fig. S12 from the Angli paper)

```{r}
color_scheme <- c("#1b9e77", "#d95f02", "#7570b3", "#e7298a", "#66a61e", 
                  "#e6ab02", "#a6761d", "#666666", "#17becf", "#bcbd22", 
                  "#ff9896", "#9467bd", "#8c564b", "#f781bf", "#8dd3c7", "#ffed6f")

```

### Male 

```{r}
M_genes <- complete_table_M %>% filter(significance == "sig") 

M_samples <- metadata %>% filter(sex == "M") %>% pull(sample_id)

# Fetch means
mean_expression_sig_genes_list <- lapply(names(table(M_genes$ct)), function(x){
  sig_genes_ct <- M_genes %>% filter(ct == x)
  colnames(mean_exps[[x]]) <- gsub("X", "", colnames(mean_exps[[x]]))
  M_samples_ct <- M_samples[M_samples %in% colnames(mean_exps[[x]])]
  mean_exps[[x]] %>% filter(row.names(.) %in% sig_genes_ct$gene) %>% select(all_of(M_samples_ct))
})

# Fetch dispersions
dispersions_sig_genes_list <- lapply(names(table(M_genes$ct)), function(x){
  sig_genes_ct <- M_genes %>% filter(ct == x)
  colnames(dispersions[[x]]) <- gsub("X", "", colnames(dispersions[[x]]))
  M_samples_ct <- M_samples[M_samples %in% colnames(dispersions[[x]])]
  dispersions[[x]] %>% filter(row.names(.) %in% sig_genes_ct$gene) %>% select(M_samples_ct)
})

# Create a list of data.frames 
estimate_per_sig_genes_list <- lapply(1:length(names(table(M_genes$ct))), function(x){
  c <- names(table(M_genes$ct))[x]
  means <- rowMeans(mean_expression_sig_genes_list[[x]], na.rm = TRUE)
  dispersion <- rowMeans(dispersions_sig_genes_list[[x]], na.rm = TRUE)
  sig_genes_ct <- M_genes %>% filter(ct == c)

  estimate_means_M <- data.frame("mean" = means, 
                                   "dispersion" = dispersion, 
                                   "signal" = ifelse(sig_genes_ct$beta < 0, "-", "+"), 
                                   "ct" = c,
                                   "gene" = names(dispersion)
                                   )
  
  row.names(estimate_means_M) <- NULL
  
  estimate_means_M

})

estimate_per_sig_genes_M <- do.call("rbind", estimate_per_sig_genes_list)


a <- ggplot(estimate_per_sig_genes_M, aes(x = log(mean, 10), y = dispersion, colour = ct)) + 
  geom_point(size = 1) +  # Adjust point size and color
  labs(
    title = "Mean vs Dispersion",    # Add plot title
    x = "AverM log10(mean expression)",                      # X-axis label
    y = "AverM dispersion"                 # Y-axis label
  ) +
  theme_minimal() +
  scale_color_manual(values = color_scheme) +
  theme(
    plot.title = element_text(hjust = 0.5),  # Center the title
    axis.title.x = element_text(margin = margin(t = 10)),  # Spacing below x-axis title
    axis.title.y = element_text(margin = margin(r = 10))   # Spacing left of y-axis title
  ) + 
  ggtitle("M") 

a

pdf(paste0(img.dir, "01_dispersions.M.pdf"), height=4, width=8)
print(a)
dev.off()
```

### Female
```{r}
F_genes <- complete_table_fe %>% filter(significance == "sig") 

F_samples <- metadata %>% filter(sex == "F") %>% pull(sample_id)

# Fetch means
mean_expression_sig_genes_list <- lapply(names(table(F_genes$ct)), function(x){
  sig_genes_ct <- F_genes %>% filter(ct == x)
  colnames(mean_exps[[x]]) <- gsub("X", "", colnames(mean_exps[[x]]))
  F_samples_ct <- F_samples[F_samples %in% colnames(mean_exps[[x]])]
  mean_exps[[x]] %>% filter(row.names(.) %in% sig_genes_ct$gene) %>% select(all_of(F_samples_ct))
})

# Fetch dispersions
dispersions_sig_genes_list <- lapply(names(table(F_genes$ct)), function(x){
  sig_genes_ct <- F_genes %>% filter(ct == x)
  colnames(dispersions[[x]]) <- gsub("X", "", colnames(dispersions[[x]]))
  F_samples_ct <- F_samples[F_samples %in% colnames(dispersions[[x]])]
  dispersions[[x]] %>% filter(row.names(.) %in% sig_genes_ct$gene) %>% select(all_of(F_samples_ct))
})

# Create a list of data.frames 
estimate_per_sig_genes_list <- lapply(1:length(names(table(F_genes$ct))), function(x){
  c <- names(table(F_genes$ct))[x]
  means <- rowMeans(mean_expression_sig_genes_list[[x]], na.rm = TRUE)
  dispersion <- rowMeans(dispersions_sig_genes_list[[x]], na.rm = TRUE)
  sig_genes_ct <- F_genes %>% filter(ct == c)

  estimate_means_fe <- data.frame("mean" = means, 
                                   "dispersion" = dispersion, 
                                   "signal" = ifelse(sig_genes_ct$beta < 0, "-", "+"), 
                                   "ct" = c,
                                   "gene" = names(dispersion)
                                   )
  
  row.names(estimate_means_fe) <- NULL
  
  estimate_means_fe

})

estimate_per_sig_genes_fe <- do.call("rbind", estimate_per_sig_genes_list)



a <- ggplot(estimate_per_sig_genes_fe, aes(x = log(mean, 10), y = dispersion, colour = ct)) + 
  geom_point(size = 1) +  # Adjust point size and color
  labs(
    title = "Mean vs Dispersion",    # Add plot title
    x = "Average log10(mean expression)",                      # X-axis label
    y = "Average dispersion"                 # Y-axis label
  ) +
  theme_minimal() +
  scale_color_manual(values = color_scheme) +
  theme(
    plot.title = element_text(hjust = 0.5),  # Center the title
    axis.title.x = element_text(margin = margin(t = 10)),  # Spacing below x-axis title
    axis.title.y = element_text(margin = margin(r = 10))   # Spacing left of y-axis title
  ) + 
  ggtitle("fe") 

a

pdf(paste0(img.dir, "01_dispersions.fe.pdf"), height=4, width=8)
print(a)
dev.off()
```


# Apply the expression filter to the data---- 

```{r}
filter_mean_ct <- c(10, 10,5,30,1.5,2,10,5,30,2,10,30,30,30,2,10)
names(filter_mean_ct) <- cts_to_study
```


## Filter using celltype specific filters

```{r}
#adjusted.pvalues are recomputed

DVG.filter.M <- lapply(names(filter_mean_ct), function(x){
  colnames(mean_exps[[x]]) <- gsub("X", "", colnames(mean_exps[[x]]))
  M_samples_ct <- M_samples[M_samples %in% colnames(mean_exps[[x]])]
  m_exp <- mean_exps[[x]] %>% select(all_of(M_samples_ct))
  mean_per_gene <- rowMeans(m_exp, na.rm = TRUE)
  th <- filter_mean_ct[x]
  genes_to_keep <- names(mean_per_gene[mean_per_gene >= th])
  DVG_cell <- DVGs$M[[x]]
  DVG_cell %>% 
    filter(gene %in% genes_to_keep) %>% 
    mutate(p_value_adj_filter = p.adjust(p_value, method = "BH"),
           significance_after_filter = ifelse(p_value_adj_filter < 0.05, "sig", "non_sig"))
})


DVG.filter.Fe <- lapply(names(filter_mean_ct), function(x){
  colnames(mean_exps[[x]]) <- gsub("X", "", colnames(mean_exps[[x]]))
  F_samples_ct <- F_samples[F_samples %in% colnames(mean_exps[[x]])]
  m_exp <- mean_exps[[x]] %>% select(all_of(F_samples_ct))
  
  mean_per_gene <- rowMeans(m_exp, na.rm = TRUE)
  th <- filter_mean_ct[x]
  genes_to_keep <- names(mean_per_gene[mean_per_gene >= th])
  DVG_cell <- DVGs$Fe[[x]]
  DVG_cell %>% 
    filter(gene %in% genes_to_keep) %>% 
    mutate(p_value_adj_filter = p.adjust(p_value, method = "BH"),
           significance_after_filter = ifelse(p_value_adj_filter < 0.05, "sig", "non_sig"))
})


names(DVG.filter.Fe) <- names(DVG.filter.M) <- names(filter_mean_ct)


DVG.filter <- list()
DVG.filter$M <- DVG.filter.M
DVG.filter$Fe <- DVG.filter.Fe


saveRDS(DVG.filter, paste0(img.dir, "/1.DVG_results.filtered.rds"))
```


## Apply the expression filter per group

Due to some specific over dispersion in some groups, specially in sex, I can apply the mean threshold per group.

```{r}
# Prepare metadata to be use
metadata_to_add <- metadata %>% 
  mutate(age_group = ifelse(age < 40, "Y", "O")) %>% 
  select(sample_id, sex, age_group)
```


```{r}

DVG.group_filter.M <- lapply(names(filter_mean_ct), function(x){
  mean_per_gene <- rowMeans(mean_exps[[x]], na.rm = TRUE)
  
  colnames(mean_exps[[x]]) <- gsub("X", "", colnames(mean_exps[[x]]))
  M_samples_ct <- M_samples[M_samples %in% colnames(mean_exps[[x]])]
  m_exp <- mean_exps[[x]] %>% select(all_of(M_samples_ct))
 
  long_df <- m_exp %>%
    mutate(gene = row.names(.)) %>%  # Add gene row numbers
    pivot_longer(cols = 1:ncol(m_exp), names_to = "sample_id", values_to = "mean") %>% 
    mutate(sample_id = gsub("X", "", sample_id)) %>%
    merge(metadata_to_add, by = "sample_id")
  
  # Summarise the dataframe by age group
  group_mean_age <- long_df %>% 
    mutate(age_group = as.factor(age_group)) %>% 
    group_by(age_group, gene) %>% 
    summarise(group_mean = mean(mean))
  
  group_mean_age_filtered <- group_mean_age %>% 
    filter(group_mean > filter_mean_ct[x])
    
  keep_genes <- names(table(group_mean_age_filtered$gene)[table(group_mean_age_filtered$gene) == 2])
  
  DVG_cell <- DVG.filter$M[[x]]
  DVG_cell %>% 
    filter(gene %in% keep_genes) %>% 
    mutate(p_value_adj_group_filter = p.adjust(p_value, method = "BH"),
           significance_after_group_filter = ifelse(p_value_adj_group_filter < 0.05, "sig", "non_sig"))
})


DVG.group_filter.F <- lapply(names(filter_mean_ct), function(x){
  mean_per_gene <- rowMeans(mean_exps[[x]], na.rm = TRUE)
  
  colnames(mean_exps[[x]]) <- gsub("X", "", colnames(mean_exps[[x]]))
  F_samples_ct <- F_samples[F_samples %in% colnames(mean_exps[[x]])]
  m_exp <- mean_exps[[x]] %>% select(all_of(F_samples_ct))
 
  long_df <- m_exp %>%
    mutate(gene = row.names(.)) %>%  # Add gene row numbers
    pivot_longer(cols = 1:ncol(m_exp), names_to = "sample_id", values_to = "mean") %>% 
    mutate(sample_id = gsub("X", "", sample_id)) %>%
    merge(metadata_to_add, by = "sample_id")
  
  # Summarise the dataframe by age group
  group_mean_age <- long_df %>% 
    mutate(age_group = as.factor(age_group)) %>% 
    group_by(age_group, gene) %>% 
    summarise(group_mean = mean(mean))
  
  group_mean_age_filtered <- group_mean_age %>% 
    filter(group_mean > filter_mean_ct[x])
    
  keep_genes <- names(table(group_mean_age_filtered$gene)[table(group_mean_age_filtered$gene) == 2])
  
  DVG_cell <- DVG.filter$Fe[[x]]
  DVG_cell %>% 
    filter(gene %in% keep_genes) %>% 
    mutate(p_value_adj_group_filter = p.adjust(p_value, method = "BH"),
           significance_after_group_filter = ifelse(p_value_adj_group_filter < 0.05, "sig", "non_sig"))
})


names(DVG.group_filter.F) <- names(DVG.group_filter.M) <- cts_to_study


DVG.group.filter <- list()
DVG.group.filter$M <- DVG.group_filter.M
DVG.group.filter$F <- DVG.group_filter.F


saveRDS(DVG.filter, paste0(img.dir, "/1.DVG.group.filter.rds"))

```

## Summarise the result

```{r}
# Male
number_of_genes_M <- lapply(names(DVG.group.filter$M), function(x){
    temp <- DVG.group.filter$M[[x]] %>% filter(significance_after_group_filter == "sig")
    data.frame("celltype" = x, "up" = sum(temp$beta > 0), "down" = sum(temp$beta < 0))
  }
)

number_of_genes_M <- do.call("rbind", number_of_genes_M)
df_melted_M <- melt(number_of_genes_M, id.vars = "celltype", variable.name = "direction", value.name = "count")


# sex 
number_of_genes_F <- lapply(names(DVG.group.filter$F), function(x){
    temp <- DVG.group.filter$F[[x]] %>% filter(significance_after_group_filter == "sig")
    data.frame("celltype" = x, "up" = sum(temp$beta > 0), "down" = sum(temp$beta < 0))
  }
)

number_of_genes_F <- do.call("rbind", number_of_genes_F)
df_melted_F <- melt(number_of_genes_F, id.vars = "celltype", variable.name = "direction", value.name = "count")

```

## Make an heatmap

```{r 01_results_after_group_filtering, fig.height=8, fig.width=7}
col = brewer.pal(9, "BuPu")[1:7]

dat1 <- number_of_genes_M[,c(2:3)]
dat2 <- number_of_genes_F[,c(2:3)]

row.names(dat1) <- number_of_genes_M[,1]
row.names(dat2) <- number_of_genes_F[,1]

#n_counts <- sample_size.filtered$N
median_cells <- sample_size$median
celltypes <- sample_size$celltype

dat1 <- dat1[celltypes, ]
dat2 <- dat2[celltypes, ]

# Create the heatmaps as you had before
row_ha_left <- rowAnnotation(
                            #`N subjects`= anno_barplot(n_counts, 
                            #                   gp = gpar(fill = "blue"), 
                            #                   width = unit(1, "cm"),  # Decreased width for the N barplot
                            #                   border = TRUE), 
                             `Median cell\nper subject` = anno_barplot(median_cells, 
                                                   gp = gpar(fill = "green"), 
                                                   width = unit(1, "cm"),  # Decreased width for the Median barplot
                                                   border = TRUE)
                             )

heatmap1 <- Heatmap(as.matrix(dat1),
                    name = "DVG",               
                    col = brewer.pal(9,"BuPu")[1:7],  
                    cluster_rows = FALSE,      
                    cluster_columns = FALSE,    
                    show_column_names = TRUE,   
                    show_row_names = FALSE,
                    left_annotation = row_ha_left,
                    column_title = "Male",
                    cell_fun = function(j, i, x, y, width, height, fill) {
                     grid.text(prettyNum(dat1[i, j], big.mark = ","), x, y, gp = gpar(fontsize = 10))}
                    )

heatmap2 <- Heatmap(as.matrix(dat2),
                    name = "DVG",               
                    col = brewer.pal(9,"BuPu")[1:7],  
                    cluster_rows = FALSE,      
                    cluster_columns = FALSE,    
                    show_column_names = TRUE,   
                    show_row_names = TRUE,
                    column_title = "Female",
                    cell_fun = function(j, i, x, y, width, height, fill) {
                     grid.text(prettyNum(dat2[i, j], big.mark = ","), x, y, gp = gpar(fontsize = 10))}
                    )





# Draw the combined heatmap with a gap between the elements
draw(heatmap1 + heatmap2 , ht_gap = unit(1, "cm"))

pdf(paste0(img.dir, "02_dvg_after_group_filtering.pdf"), height=6, width=5)
draw(heatmap1 + heatmap2, ht_gap = unit(1, "cm"))
dev.off()
```

## Plot mean vs dispersion in significant genes (Fig. S12 from the Angli paper)

```{r}
#color_scheme <- c("#1f77b4", "#ff7f0e", "#2ca02c", "#d62728") 
```

```{r}
complete_table_male.group_filter <- do.call("rbind", lapply(names(DVG.group.filter$M), function(x) DVG.group.filter$M[[x]] %>% mutate(ct = x)))
complete_table_female.group_filter <- do.call("rbind", lapply(names(DVG.group.filter$F), function(x) DVG.group.filter$F[[x]] %>% mutate(ct = x)))

```


### Male

```{r}
m_genes <- complete_table_male.group_filter %>% filter(significance_after_filter == "sig") 

# Fetch means
mean_expression_sig_genes_list <- lapply(names(table(m_genes$ct)), function(x){
  sig_genes_ct <- m_genes %>% filter(ct == x)
  colnames(mean_exps[[x]]) <- gsub("X", "", colnames(mean_exps[[x]]))
  M_samples_ct <- M_samples[M_samples %in% colnames(mean_exps[[x]])]
  m_exp <- mean_exps[[x]] %>% select(all_of(M_samples_ct))
  
  m_exp %>% filter(row.names(.) %in% sig_genes_ct$gene)
})



# Fetch dispersions
dispersions_sig_genes_list <- lapply(names(table(m_genes$ct)), function(x){
  sig_genes_ct <- m_genes %>% filter(ct == x)
  colnames(dispersions[[x]]) <- gsub("X", "", colnames(dispersions[[x]]))
  M_samples_ct <- M_samples[M_samples %in% colnames(dispersions[[x]])]
  m_dist <- dispersions[[x]] %>% select(all_of(M_samples_ct))
  
  m_dist %>% filter(row.names(.) %in% sig_genes_ct$gene)
})

# Create a list of data.frames 
estimate_per_sig_genes_list <- lapply(1:length(names(table(m_genes$ct))), function(x){
  c <- names(table(m_genes$ct))[x]
  means <- rowMeans(mean_expression_sig_genes_list[[x]], na.rm = TRUE)
  dispersion <- rowMeans(dispersions_sig_genes_list[[x]], na.rm = TRUE)
  sig_genes_ct <- m_genes %>% filter(ct == c)

  estimate_means_M <- data.frame("mean" = means, 
                                   "dispersion" = dispersion, 
                                   "signal" = ifelse(sig_genes_ct$beta < 0, "-", "+"), 
                                   "ct" = c,
                                   "gene" = names(dispersion)
                                   )
  
  row.names(estimate_means_M) <- NULL
  
  estimate_means_M

})

estimate_per_sig_genes_M <- do.call("rbind", estimate_per_sig_genes_list)



a <- ggplot(estimate_per_sig_genes_M, aes(x = log(mean, 10), y = dispersion, colour = ct)) + 
  geom_point(size = 2) +  # Adjust point size and color
  labs(
    title = "Mean vs Dispersion",    # Add plot title
    x = "Average log10(mean expression)",                      # X-axis label
    y = "Average dispersion"                 # Y-axis label
  ) +
  theme_minimal() +
  #scale_color_manual(values = color_scheme) +
  theme(
    plot.title = element_text(hjust = 0.5),  # Center the title
    axis.title.x = element_text(margin = margin(t = 10)),  # Spacing below x-axis title
    axis.title.y = element_text(margin = margin(r = 10))   # Spacing left of y-axis title
  ) + 
  ggtitle("Male") 

a

pdf(paste0(img.dir, "02_dispersions.M.pdf"), height=4, width=8)
print(a)
dev.off()
```
### Female

```{r}
f_genes <- complete_table_female.group_filter %>% filter(significance_after_filter == "sig") 

# Fetch means
mean_expression_sig_genes_list <- lapply(names(table(f_genes$ct)), function(x){
  sig_genes_ct <- f_genes %>% filter(ct == x)
  colnames(mean_exps[[x]]) <- gsub("X", "", colnames(mean_exps[[x]]))
  F_samples_ct <- F_samples[F_samples %in% colnames(mean_exps[[x]])]
  m_exp <- mean_exps[[x]] %>% select(all_of(F_samples_ct))
  
  m_exp %>% filter(row.names(.) %in% sig_genes_ct$gene)
})


# Fetch dispersions
dispersions_sig_genes_list <- lapply(names(table(f_genes$ct)), function(x){
  sig_genes_ct <- f_genes %>% filter(ct == x)
  colnames(dispersions[[x]]) <- gsub("X", "", colnames(dispersions[[x]]))
  F_samples_ct <- F_samples[F_samples %in% colnames(dispersions[[x]])]
  f_dist <- dispersions[[x]] %>% select(all_of(F_samples_ct))
  
  f_dist %>% filter(row.names(.) %in% sig_genes_ct$gene)
})

# Create a list of data.frames 
estimate_per_sig_genes_list <- lapply(1:length(names(table(f_genes$ct))), function(x){
  c <- names(table(f_genes$ct))[x]
  means <- rowMeans(mean_expression_sig_genes_list[[x]], na.rm = TRUE)
  dispersion <- rowMeans(dispersions_sig_genes_list[[x]], na.rm = TRUE)
  sig_genes_ct <- f_genes %>% filter(ct == c)

  estimate_means_F <- data.frame("mean" = means, 
                                   "dispersion" = dispersion, 
                                   "signal" = ifelse(sig_genes_ct$beta < 0, "-", "+"), 
                                   "ct" = c,
                                   "gene" = names(dispersion)
                                   )
  
  row.names(estimate_means_F) <- NULL
  
  estimate_means_F

})

estimate_per_sig_genes_F <- do.call("rbind", estimate_per_sig_genes_list)



a <- ggplot(estimate_per_sig_genes_F, aes(x = log(mean, 10), y = dispersion, colour = ct)) + 
  geom_point(size = 2) +  # Adjust point size and color
  labs(
    title = "Mean vs Dispersion",    # Add plot title
    x = "Average log10(mean expression)",                      # X-axis label
    y = "Average dispersion"                 # Y-axis label
  ) +
  theme_minimal() +
  #scale_color_manual(values = color_scheme) +
  theme(
    plot.title = element_text(hjust = 0.5),  # Center the title
    axis.title.x = element_text(margin = margin(t = 10)),  # Spacing below x-axis title
    axis.title.y = element_text(margin = margin(r = 10))   # Spacing left of y-axis title
  ) + 
  ggtitle("Female") 

a

pdf(paste0(img.dir, "02_dispersions.F.pdf"), height=4, width=8)
print(a)
dev.off()
```

# Compare per cell type 

```{r}
overlap_do <- data.frame("up" = c(), "down" = c(), "ct" = c())

for (ct in names(DVG.filter$M)){
  male_genes_up <- DVG.filter$M[[ct]] %>% filter(significance_after_filter == "sig" & beta > 0) %>% pull(gene)
  male_genes_down <- DVG.filter$M[[ct]] %>% filter(significance_after_filter == "sig" & beta < 0) %>% pull(gene)
  
  female_genes_up <- DVG.filter$Fe[[ct]] %>% filter(significance_after_filter == "sig" & beta > 0) %>% pull(gene)
  female_genes_down <- DVG.filter$Fe[[ct]] %>% filter(significance_after_filter == "sig" & beta < 0) %>% pull(gene)
 
  up <- length(intersect(male_genes_up, female_genes_up))
  down <- length(intersect(male_genes_down, female_genes_down))
  
  overlap_do <- rbind(overlap_do, data.frame(up, down, ct))
}

row.names(overlap_do) <- overlap_do$ct
overlap_do$ct <- NULL

heatmap1 <- Heatmap(as.matrix(overlap_do),
                    name = "DVG in Common",               
                    col = brewer.pal(9,"BuPu")[1:7],  
                    cluster_rows = FALSE,      
                    cluster_columns = FALSE,    
                    show_column_names = TRUE,   
                    show_row_names = FALSE,
                    #left_annotation = row_ha_left,
                    column_title = "",
                    cell_fun = function(j, i, x, y, width, height, fill) {
                     grid.text(prettyNum(overlap_do[i, j], big.mark = ","), x, y, gp = gpar(fontsize = 10))}
                    )

pdf(paste0(img.dir, "03_common_genes.pdf"), height=4, width=3)
print(heatmap1)
dev.off()
```
